[Tests] split out db initialization during tests
- instead of initing all dbs every test.
