+++ lib/src/main/kotlin/com/dbflow5/database/DatabaseHelperDelegate.kt
@@ -125,6 +125,7 @@ class DatabaseHelperDelegate(
                 context.assets.open(prepackagedName)
             }
             writeDB(dbPath, inputStream)
+            databaseDefinition.reopen()
         } catch (e: IOException) {
             FlowLog.log(FlowLog.Level.W, "Failed to open file", e)
         }
