+++ tests/src/androidTest/java/com/dbflow5/User.kt
@@ -3,8 +3,9 @@ package com.dbflow5
 import com.dbflow5.annotation.Column
 import com.dbflow5.annotation.PrimaryKey
 import com.dbflow5.annotation.Table
+import com.dbflow5.contentobserver.ContentObserverDatabase
 
-@Table(database = AppDatabase::class, name = "User2")
+@Table(database = ContentObserverDatabase::class, name = "User2")
 class User(@PrimaryKey var id: Int = 0,
            @Column var firstName: String? = null,
            @Column var lastName: String? = null,
