+++ tests/src/androidTest/java/com/dbflow5/contentobserver/ContentObserverDatabase.kt
@@ -1,7 +1,7 @@
-package com.dbflow5
+package com.dbflow5.contentobserver
 
 import com.dbflow5.annotation.Database
 import com.dbflow5.config.DBFlowDatabase
 
 @Database(version = 1)
-abstract class AppDatabase : DBFlowDatabase()
+abstract class ContentObserverDatabase : DBFlowDatabase()
