+++ tests/src/androidTest/java/com/dbflow5/contentobserver/ContentObserverTest.kt
@@ -1,30 +1,44 @@
 package com.dbflow5.contentobserver
 
 import android.net.Uri
-import com.dbflow5.BaseUnitTest
+import androidx.test.core.app.ApplicationProvider
+import com.dbflow5.DBFlowInstrumentedTestRule
 import com.dbflow5.DemoApp
+import com.dbflow5.ImmediateTransactionManager
 import com.dbflow5.TABLE_QUERY_PARAM
+import com.dbflow5.config.DBFlowDatabase
+import com.dbflow5.config.database
 import com.dbflow5.config.databaseForTable
 import com.dbflow5.config.modelAdapter
 import com.dbflow5.config.tableName
 import com.dbflow5.contentobserver.User_Table.id
 import com.dbflow5.contentobserver.User_Table.name
+import com.dbflow5.database.AndroidSQLiteOpenHelper
 import com.dbflow5.database.DatabaseWrapper
 import com.dbflow5.getNotificationUri
 import com.dbflow5.query.SQLOperator
 import com.dbflow5.query.delete
+import com.dbflow5.runtime.ContentResolverNotifier
 import com.dbflow5.runtime.FlowContentObserver
 import com.dbflow5.structure.ChangeAction
-import com.dbflow5.structure.delete
-import com.dbflow5.structure.insert
-import com.dbflow5.structure.save
-import com.dbflow5.structure.update
 import org.junit.Assert.assertEquals
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
 import java.util.concurrent.CountDownLatch
 
-class ContentObserverTest : BaseUnitTest() {
+class ContentObserverTest {
+
+    @JvmField
+    @Rule
+    var dblflowTestRule = DBFlowInstrumentedTestRule.create {
+        database<ContentObserverDatabase>({
+            modelNotifier(ContentResolverNotifier(DemoApp.context, "com.grosner.content"))
+            transactionManagerCreator { databaseDefinition: DBFlowDatabase ->
+                ImmediateTransactionManager(databaseDefinition)
+            }
+        }, AndroidSQLiteOpenHelper.createHelperCreator(ApplicationProvider.getApplicationContext()))
+    }
 
     val contentUri = "com.grosner.content"
 
@@ -32,7 +46,7 @@ class ContentObserverTest : BaseUnitTest() {
 
     @Before
     fun setupUser() {
-        databaseForTable<User> { dbFlowDatabase ->
+        database<ContentObserverDatabase> { dbFlowDatabase ->
             delete<User>().execute(dbFlowDatabase)
         }
         user = User(5, "Something", 55)
@@ -60,7 +74,7 @@ class ContentObserverTest : BaseUnitTest() {
 
     @Test
     fun testSpecificUrlUpdate() {
-       // assertProperConditions(ChangeAction.UPDATE) { user, db -> user.apply { age = 56 }.update(db) }
+        // assertProperConditions(ChangeAction.UPDATE) { user, db -> user.apply { age = 56 }.update(db) }
 
     }
 
@@ -72,8 +86,8 @@ class ContentObserverTest : BaseUnitTest() {
 
     @Test
     fun testSpecificUrlDelete() {
-       // user.save(databaseForTable<User>())
-       // assertProperConditions(ChangeAction.DELETE) { user, db -> user.delete(db) }
+        // user.save(databaseForTable<User>())
+        // assertProperConditions(ChangeAction.DELETE) { user, db -> user.delete(db) }
     }
 
     private fun assertProperConditions(action: ChangeAction, userFunc: (User, DatabaseWrapper) -> Unit) {
